feat: implement shared build runner and reorganize CLI structure
- Add shared build runner (internal/build/runner.go) to deduplicate build orchestration
  between CLI and daemon
- Wire daemon to use runner for real builds with log streaming
- Refactor CLI build command to use shared runner for normal builds
- Preserve legacy CLI path for test-mode/fetch-only to maintain CI compatibility
- Add customer field to StartBuildRequest proto for build ID scoping
- Implement LoadFromBytes in config package for inline config support
- Fix daemon log buffer to use pointers to avoid copying mutex types

- Reorganize internal/cli into subpackages for better structure:
  - internal/cli/build/ - build command and related logic
  - internal/cli/client/ - all client/daemon interaction commands
- Update .gitignore to only ignore top-level /build/ directory

All tests pass and CLI functionality verified.
